test: FavoriteRestController 테스트

+++ src/main/java/com/project/Project/controller/interaction/controller/FavoriteRestController.java
@@ -7,31 +7,36 @@ import com.project.Project.controller.interaction.dto.FavoriteResponseDto;
 import com.project.Project.domain.Member;
 import com.project.Project.domain.building.Building;
 import com.project.Project.domain.enums.MemberRole;
+import com.project.Project.exception.ErrorCode;
+import com.project.Project.exception.building.BuildingException;
+import com.project.Project.exception.interaction.FavoriteException;
 import com.project.Project.serializer.building.BuildingSerializer;
 import com.project.Project.service.FavoriteService;
 import com.project.Project.validator.ExistBuilding;
-import com.project.Project.validator.interaction.ExistFavorite;
-import com.project.Project.validator.interaction.FavoriteExistVO;
+
 import com.project.Project.validator.interaction.FavoriteExistValidator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
-
+@Validated
 @RestController
 @RequiredArgsConstructor
 public class FavoriteRestController {
     // todo : 사용자 정보 파라미터 모든 메소드에 필요
 
     private final FavoriteService favoriteService;
+    private final FavoriteExistValidator favoriteExistValidator;
 
     private Member getTestMember() {
          return  Member.builder() // temp user
@@ -54,7 +59,7 @@ public class FavoriteRestController {
         return ResponseEntity.ok(QueryDslUtil.toSlice(buildingListResponse, pageable));
     }
 
-    @PostMapping("/member/favorite")
+    @PostMapping("/member/favorite/{buildingId}")
     public ResponseEntity<FavoriteResponseDto.FavoriteAddResponse> addFavoriteBuilding(@ExistBuilding Long buildingId) {
         Long savedFavoriteId = favoriteService.addFavoriteBuilding(buildingId, getTestMember());
         return ResponseEntity.ok(FavoriteResponseDto.FavoriteAddResponse.builder()
@@ -64,11 +69,11 @@ public class FavoriteRestController {
         );
     }
 
-    @DeleteMapping("/member/favorite")
+    @DeleteMapping("/member/favorite/{buildingId}")
     public ResponseEntity<FavoriteResponseDto.FavoriteDeleteResponse> deleteFavoriteBuilding(@ExistBuilding Long buildingId) {
         Member member = getTestMember();
-        @ExistFavorite
-        FavoriteExistVO favoriteExistVO = FavoriteExistVO.builder().member(member).buildingId(buildingId).build();
+        if(!favoriteExistValidator.isValid(member, buildingId))
+            throw new FavoriteException(ErrorCode.FAVORITE_NOT_FOUND);
 
         Long deletedFavoriteId = favoriteService.deleteFavoriteBuilding(buildingId, member);
         return ResponseEntity.ok(FavoriteResponseDto.FavoriteDeleteResponse.builder()
